fancy shit and trails

+++ src/ball.h
@@ -1,6 +1,6 @@
 typedef struct {
     int x;
     int y;
-    int speedx;
-    int speedy;
+    float speedx;
+    float speedy;
 } Ball;
